adding an htmldoc filter
Use it to create a pdf from html content. In this example for the user
manual an index.html as well as an indes.pdf will be created.

<pre>
---
title:      User Manual
created_at: 2007-08-29 08:57:00.000000 -06:00
filter:
  - erb
  - textile
  - outline
---
extension: pdf
filter:
  - erb
  - textile
  - outline
  - htmldoc: no_toc, landscape
---
</pre>

Note that no additional layout is needed. It's pulled in just before the
htmldoc filter get's the input. This allows creating pdf for individual
pages using the site wide html layout without duplicating it only for the
htmldoc filter.