changed @page.latex to @page.format == :latex
This approach is much more flexible and extensible to other formats. The
default format is :html, but it can be overridden by
SITE.page_defaults['format'] and by the individual pages meta data "format: latex"
Many filters and helpers that can produce multiple output formats are sensitive
to this option.